requirements.txt
PyGithub==1.55
requests==2.25.1
pytest==7.1.2